Struct vec1::Vec1

pub struct Vec1<I>(_);

std::vec::Vec wrapper which guarantees to have at least 1 element.

Vec1<T> dereferences to &[T] and &mut [T] as functionality exposed through this can not change the length.

Methods of Vec which can be called without reducing the length (e.g. capacity(), reserve()) are exposed through wrappers with the same function signature.

Methods of Vec which could reduce the length to 0 are return a Result wrapping their normal return type.

Methods with returned Option<T> with None if the length was 0 (and do not reduce the length) now return T. (e.g. first, last, first_mut, etc.).

All stable traits and methods implemented on Vec<T> should also be implemented on Vec1<T> (except if they make no sense to implement due to the len 1 guarantee). Be aware implementations may lack behind a bit, fell free to open a issue/make a PR, but please search closed and open issues for duplicates first.

Implementations

impl<I> Vec1<I>

pub fn new(first: I) -> Self

Creates a new instance containing a single element.

pub fn with_capacity(first: I, capacity: usize) -> Self

Creates a new instance with a given capacity and a given “first” element.

pub fn try_from_vec(vec: Vec<I>) -> Result<Self, Size0Error>

Tries to create a instance from a normal Vec<T>.

Errors

This will fail if the input Vec<T> is empty. The returned error is a Size0Error instance, as such this means the input vector will be dropped if it’s empty. But this is normally fine as it only happens if the Vec<T> is empty.

pub fn last(&self) -> &I

Returns a reference to the last element.

As $name always contains at least one element there is always a last element.

pub fn last_mut(&mut self) -> &mut I

Returns a mutable reference to the last element.

As $name always contains at least one element there is always a last element.

pub fn first(&self) -> &I

Returns a reference to the first element.

As $name always contains at least one element there is always a first element.

pub fn first_mut(&mut self) -> &mut I

Returns a mutable reference to the first element.

As $name always contains at least one element there is always a first element.

pub fn truncate(&mut self, len: usize) -> Result<(), Size0Error>

Truncates the SmalVec1 to given length.

Errors

If len is 0 an error is returned as the length >= 1 constraint must be uphold.

pub fn try_truncate(&mut self, len: usize) -> Result<(), Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use truncate

Truncates the SmalVec1 to given length.

Errors

If len is 0 an error is returned as the length >= 1 constraint must be uphold.

pub fn swap_remove(&mut self, index: usize) -> Result<I, Size0Error>

Calls swap_remove on the inner smallvec if length >= 2.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

pub fn try_swap_remove(&mut self, index: usize) -> Result<I, Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use swap_remove

Calls swap_remove on the inner smallvec if length >= 2.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

pub fn remove(&mut self, index: usize) -> Result<I, Size0Error>

Calls remove on the inner smallvec if length >= 2.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

pub fn try_remove(&mut self, index: usize) -> Result<I, Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use remove, also try_remove PANICS on out of bounds

Calls remove on the inner smallvec if length >= 2.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

Panics

If index is greater or equal then len.

pub fn drain<R>(&mut self, range: R) -> Result<Drain<'_, I>, Size0Error> where
R: RangeBounds<usize>,

If calls drain on the underlying vector if it will not empty the vector.

Error

If calling drain would empty the vector an Err(Size0Error) is returned instead of draining the vector.

Panic

Like Vec::drain() panics if:

The starting point is greater than the end point.

The end point is greater than the length of the vector.

pub fn retain<F>(&mut self, f: F) -> Result<(), Size0Error> where
F: FnMut(&I) -> bool,

Removes all elements except the ones which the predicate says need to be retained.

The moment the last element would be removed this will instead fail, not removing the element. All but the last element will have been removed anyway.

Error

If the last element would be removed instead of removing it a Size0Error is returned.

Example

Is for Vec1 but similar code works with SmallVec1, too.

let mut vec = vec1![1, 7, 8, 9, 10];
vec.retain(|v| *v % 2 == 1).unwrap();
assert_eq!(vec, vec1![1, 7, 9]);
let Size0Error = vec.retain(|_| false).unwrap_err();
assert_eq!(vec.len(), 1);
assert_eq!(vec.last(), &9);

pub fn dedup_by_key<F, K>(&mut self, key: F) where
F: FnMut(&mut I) -> K,
K: PartialEq<K>,

Calls dedup_by_key on the inner smallvec.

While this can remove elements it will never produce a empty vector from an non empty vector.

pub fn dedup_by<F>(&mut self, same_bucket: F) where
F: FnMut(&mut I, &mut I) -> bool,

Calls dedup_by_key on the inner smallvec.

While this can remove elements it will never produce a empty vector from an non empty vector.

pub fn pop(&mut self) -> Result<I, Size0Error>

Remove the last element from this vector, if there is more than one element in it.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

pub fn try_pop(&mut self) -> Result<I, Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use pop

Remove the last element from this vector, if there is more than one element in it.

Errors

If len is 1 an error is returned as the length >= 1 constraint must be uphold.

pub fn resize_with<F>(&mut self, new_len: usize, f: F) -> Result<(), Size0Error> where
F: FnMut() -> I,

See Vec::resize_with() but fails if it would resize to length 0.

pub fn try_resize_with<F>(
&mut self,
new_len: usize,
f: F
) -> Result<(), Size0Error> where
F: FnMut() -> I,

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use resize_with

See Vec::resize_with() but fails if it would resize to length 0.

pub fn split_off_first(self) -> (I, Vec<I>)

Splits off the first element of this vector and returns it together with the rest of the vector.

pub fn split_off_last(self) -> (Vec<I>, I)

Splits off the last element of this vector and returns it together with the rest of the vector.

pub fn into_boxed_slice(self) -> Box<[I]>

Turns this vector into a boxed slice.

For Vec1 this is as cheap as for Vec but for SmallVec1 this will cause an allocation if the on-stack buffer was not yet spilled.

pub fn leak<'a>(self) -> &'a mut [I] where
I: 'a,

Leaks the allocation to return a mutable slice reference.

This is equivalent to turning this vector into a boxed slice and then leaking that slice.

In case of SmallVec1 calling leak does entail an allocation if the stack-buffer had not yet spilled.

impl<I> Vec1<I>

pub fn append(&mut self, other: &mut Vec<I>)

See Vec for a rough idea how this method works.

pub fn reserve(&mut self, additional: usize)

See Vec for a rough idea how this method works.

pub fn reserve_exact(&mut self, additional: usize)

See Vec for a rough idea how this method works.

pub fn shrink_to_fit(&mut self)

See Vec for a rough idea how this method works.

pub fn as_mut_slice(&mut self) -> &mut [I]

See Vec for a rough idea how this method works.

pub fn push(&mut self, value: I)

See Vec for a rough idea how this method works.

pub fn insert(&mut self, idx: usize, val: I)

See Vec for a rough idea how this method works.

pub fn len(&self) -> usize

See Vec for a rough idea how this method works.

pub fn capacity(&self) -> usize

See Vec for a rough idea how this method works.

pub fn as_slice(&self) -> &[I]

See Vec for a rough idea how this method works.

impl<I> Vec1<I> where
I: PartialEq<I>,

pub fn dedup(&mut self)

impl<I> Vec1<I> where
I: Copy,

pub fn extend_from_slice(&mut self, slice: &[I])

impl<I> Vec1<I> where
I: Clone,

pub fn resize(&mut self, len: usize, value: I) -> Result<(), Size0Error>

See Vec::resize() but fails if it would resize to length 0.

pub fn try_resize(&mut self, len: usize, value: I) -> Result<(), Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use resize_with

See Vec::resize() but fails if it would resize to length 0.

impl<T> Vec1<T>

pub fn from_vec(vec: Vec<T>) -> StdResult<Self, Vec<T>>

👎 Deprecated since 1.2.0:
does not work with ? use Vec1::try_from_vec() instead

Tries to create a Vec1<T> from a Vec<T>.

The fact that the input is returned as error if it’s empty, means that it doesn’t work well with the ? operator. It naming is also semantic sub-optimal as it’s not a “from” but “try from” conversion. Which is why this method is now deprecated. Instead use try_from_vec and once TryFrom is stable it will be possible to use try_from, too.

Errors

If the input is empty the input is returned as error.

pub fn into_vec(self) -> Vec<T>

Turns this Vec1 into a Vec.

pub fn as_vec(&self) -> &Vec<T>

Return a reference to the underlying Vec.

pub fn mapped<F, N>(self, map_fn: F) -> Vec1<N> where
F: FnMut(T) -> N,

Create a new Vec1 by consuming self and mapping each element.

This is useful as it keeps the knowledge that the length is >= 1, even through the old Vec1 is consumed and turned into an iterator.

Example

let data = vec1![1u8,2,3];

let data = data.mapped(|x|x*2);
assert_eq!(data, vec![2,4,6]);

// without mapped
let data = Vec1::try_from_vec(data.into_iter().map(|x|x*2).collect::<Vec<_>>()).unwrap();
assert_eq!(data, vec![4,8,12]);

pub fn mapped_ref<F, N>(&self, map_fn: F) -> Vec1<N> where
F: FnMut(&T) -> N,

Create a new Vec1 by mapping references to the elements of self.

The benefit to this compared to Iterator::map is that it’s known that the length will still be at least 1 when creating the new Vec1.

pub fn mapped_mut<F, N>(&mut self, map_fn: F) -> Vec1<N> where
F: FnMut(&mut T) -> N,

Create a new Vec1 by mapping mutable references to the elements of self.

The benefit to this compared to Iterator::map is that it’s known that the length will still be at least 1 when creating the new Vec1.

pub fn try_mapped<F, N, E>(self, map_fn: F) -> Result<Vec1<N>, E> where
F: FnMut(T) -> Result<N, E>,

Create a new Vec1 by consuming self and mapping each element to a Result.

This is useful as it keeps the knowledge that the length is >= 1, even through the old Vec1 is consumed and turned into an iterator.

As this method consumes self, returning an error means that this vec is dropped. I.e. this method behaves roughly like using a chain of into_iter(), map, collect::<Result<Vec<N>,E>> and then converting the Vec back to a Vec1.

Errors

Once any call to map_fn returns a error that error is directly returned by this method.

Example

let data = vec1![1,2,3];

let data: Result<Vec1<u8>, &'static str> = data.try_mapped(|x| Err("failed"));
assert_eq!(data, Err("failed"));

pub fn try_mapped_ref<F, N, E>(&self, map_fn: F) -> Result<Vec1<N>, E> where
F: FnMut(&T) -> Result<N, E>,

Create a new Vec1 by mapping references to the elements of self to Results.

The benefit to this compared to Iterator::map is that it’s known that the length will still be at least 1 when creating the new Vec1.

Errors

Once any call to map_fn returns a error that error is directly returned by this method.

pub fn try_mapped_mut<F, N, E>(&mut self, map_fn: F) -> Result<Vec1<N>, E> where
F: FnMut(&mut T) -> Result<N, E>,

Create a new Vec1 by mapping mutable references to the elements of self to Results.

The benefit to this compared to Iterator::map is that it’s known that the length will still be at least 1 when creating the new Vec1.

Errors

Once any call to map_fn returns a error that error is directly returned by this method.

pub fn split_off(&mut self, at: usize) -> Result<Vec1<T>, Size0Error>

Class split_off on the wrapped vector

Panics

If at is greater then len. (In the same way [Vec.split_off()] does.)

Errors

If splitting would result in an empty Vec1 an error is returned, this happens if at is 0 or at is equals to len.

pub fn try_split_off(&mut self, at: usize) -> Result<Vec1<T>, Size0Error>

👎 Deprecated since 1.8.0:
try_ prefix created ambiguity use split_off, try_split_off doesn't panic on out of bounds at`

Calls split_off on the inner vec if both resulting parts have length >= 1.

In difference to split_off this also returns a Size0Error if at is greater then len. Which is different to how [Vec.split_off()] behaves.

Errors

If after the split any part would be empty an error is returned as the length >= 1 constraint must be uphold.

pub fn splice<R, I>(
&mut self,
range: R,
replace_with: I
) -> Result<Splice<'_, <I as IntoIterator>::IntoIter>, Size0Error> where
I: IntoIterator<Item = T>,
R: RangeBounds<usize>,

Calls splice on the underlying vec (only) if it wont produce an empty vec.

Errors

If range covers the whole vec and the replacement iterator doesn’t yield any value an error is returned instead of doing any splicing.

To check if the iterator will yield values we need to turn call next on it once which means that if an error is returned Iterator::next() is still called once!

Panics

This will panic under the same conditions as Vec::splice(), the conditions are:

if the starting point is greater than the end point

if the end point is greater than the length of the vector.

impl Vec1<u8>

pub fn to_ascii_uppercase(&self) -> Vec1<u8>

Works like &[u8].to_ascii_uppercase() but returns a Vec1<T> instead of a Vec<T>

pub fn to_ascii_lowercase(&self) -> Vec1<u8>

Works like &[u8].to_ascii_lowercase() but returns a Vec1<T> instead of a Vec<T>